preprocess-service/requirements.txt
fastapi
uvicorn[standard]
minio
pydantic
python-dotenv
pydantic-settings
Pillow
opencv-python-headless
imageio[ffmpeg]
